+++ apps/mobile/src/app/(home,search,profile,notifications,settings)/search.tsx
@@ -166,7 +166,7 @@ const stylesheet = createStyleSheet((theme) => ({
     width: theme.space[7],
   },
   query: {
-    backgroundColor: theme.colors.gray.ui,
+    backgroundColor: theme.colors.gray.uiActive,
     borderWidth: 0,
   },
 }))